Spatial Cross-Validation

Spatial cross-validation splits the data into V groups of disjointed sets. These sets are separated using spatial coordinates such that observations within the set are geographically closer to each other. This is helpful to reduce spatial autocorrelation compared to random cross validation.

Usage

spatial_block_cv(
  data,
  coords,
  v = 10,
  nrow = 5,
  ncol = 5,
  crs = 4326,
  outcome = "",
  allocation = c("random", "systematic", "checkboard"),
  save_blocks = FALSE,
  verbose = FALSE,
  progress = FALSE,
  show_blocks = FALSE,
  ...
)

spatial_clustering_cv(
  data,
  coords,
  v = 10,
  method = c("kmeans", "kmedoids"),
  ...
)

Arguments

data

A data frame.

coords

A vector of variable names, typically spatial coordinates, to partition the data into disjointed sets via k-means clustering.

v

The number of partitions of the data set.

...

Extra arguments passed on to clustering functions, stats::kmeans() for kmeans, and.

method

A character string for clustering algorithm (i.e., "kmeans" or "kmedoids")

Value

A tibble with classes spatial_cv, rset, tbl_df, tbl, and data.frame. The results include a column for the data split objects and an identification variable id.

Details

The variables in the coords argument are used for k-means clustering of the data into disjointed sets, as outlined in Brenning (2012). These clusters are used as the folds for cross-validation. Depending on how the data are distributed spatially, there may not be an equal number of points in each fold.

spatial_clustering_cv uses clustering algorithms to create spatial clusters. A resample of the analysis data consists of V-1 of the folds/clusters while the assessment set contains the final fold/cluster. In basic spatial cross-validation (i.e. no repeats), the number of resamples is equal to V.

spatial_block_cv makes partitions with spatial blocks or buffers and allocate them into cross validation folds. Blocks are units of geographical area (e.g. rectangles, spatial polygons and buffers of specific distance) Within these units, all species data are treated together—for instance, allocated to the same fold of cv. Several blocks could be allocated to one cv fold.

References

Valavi, R, Elith, J, Lahoz-Monfort, JJ, Guillera-Arroita, G. blockCV: An r package for generating spatially or environmentally separated folds for k-fold cross-validation of species distribution models. Methods Ecol Evol. 2019; 10: 225– 232. https://doi.org/10.1111/2041-210X.13107

A. Brenning, "Spatial cross-validation and bootstrap for the assessment of prediction rules in remote sensing: The R package sperrorest," 2012 IEEE International Geoscience and Remote Sensing Symposium, Munich, 2012, pp. 5372-5375, doi: 10.1109/IGARSS.2012.6352393.

Examples

data(ames, package = "modeldata")
spatial_clustering_cv(ames, coords = c(Latitude, Longitude), v = 5)
#> #  5-fold spatial clustering cross-validation 
#> # A tibble: 5 × 2
#>   splits             id   
#>   <list>             <chr>
#> 1 <split [2069/861]> Fold1
#> 2 <split [2767/163]> Fold2
#> 3 <split [2283/647]> Fold3
#> 4 <split [2040/890]> Fold4
#> 5 <split [2561/369]> Fold5